Replace x.com with twittter.com, and prevent sys.exit
The script to download twitter videos was written before x.com, so cleanup the url before passing it to the library.

Furthermore, the library calls exit() on errors in places instead of properly raising a normal exception, so we have to separately catch SystemExit (which is a BaseException, not an Exception) to prevent that.