Defer discord message before processing results
Discord expects a response within 3s on an interaction, any longer and we get a "Unknown interaction" error. Since it can take 5-10s to process results, and even longer if we are in a queue, we defer before we start processing the results, and send the response as a follow up instead.